move mock request to mock directory
